feat: update md file (sample rate) + linter
